src/app/feedback/feedback.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';
import { ReactiveFormsModule } from '@angular/forms';
import { MatInputModule, MatSelectModule, MatRadioModule } from '@angular/material';
import { BrowserAnimationsModule } from '@angular/platform-browser/animations';

import { FeedbackComponent } from './feedback.component';
import { UsersurveyComponent } from '../usersurvey/usersurvey.component';

describe('FeedbackComponent', () => {
  let component: FeedbackComponent;
  let fixture: ComponentFixture<FeedbackComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ FeedbackComponent, UsersurveyComponent ],
      imports: [ReactiveFormsModule, MatInputModule, MatSelectModule, MatRadioModule, BrowserAnimationsModule ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(FeedbackComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('submit should work', () => {
    expect(component.submitted).toBe(false);
    component.onSubmit();
    expect(component.submitted).toBe(true);
  });
});
